Se borran componentes innecesarios y se cambia la funcion savepost al archivo de CRUDposts
