+++ src/index.html
@@ -40,9 +40,9 @@
     const auth = firebase.auth();
     const fs = firebase.firestore();
   </script>
-  
+
   <script type="module" src="app.js"></script>
-  <script defer src="./lib/index.js"></script>
+  <!--<script defer src="./lib/index.js"></script>-->
 
 </body>
